+++ apps/cowswap-frontend/src/common/hooks/useMultipleOrdersCancellation/useCancelMultipleOrders.ts
@@ -1,5 +1,6 @@
 import { useCallback } from 'react'
 
+import { COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS, isBarnBackendEnv } from '@cowprotocol/common-utils'
 import { OrderSigningUtils } from '@cowprotocol/cow-sdk'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
@@ -26,7 +27,10 @@ export function useCancelMultipleOrders(): (orders: CancellableOrder[]) => Promi
       }
 
       const orderUids = ordersToCancel.map((order) => order.id)
-      const { signature, signingScheme } = await OrderSigningUtils.signOrderCancellations(orderUids, chainId, signer)
+      const { signature, signingScheme } = await OrderSigningUtils.signOrderCancellations(orderUids, chainId, signer, {
+        env: isBarnBackendEnv ? 'staging' : 'prod',
+        settlementContractOverride: COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS,
+      })
 
       if (!signature) throw new Error(t`Signature is undefined!`)