+++ apps/cowswap-frontend/src/modules/account/containers/CowShedInfo/index.tsx
@@ -1,6 +1,7 @@
 import { ReactNode } from 'react'
 
 import { ACCOUNT_PROXY_LABEL } from '@cowprotocol/common-const'
+import { isInjectedWidget } from '@cowprotocol/common-utils'
 import { Command } from '@cowprotocol/types'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
@@ -9,7 +10,8 @@ import { Pocket } from 'react-feather'
 import { Link } from 'react-router'
 import styled from 'styled-components/macro'
 
-import { getProxyAccountUrl, useCurrentAccountProxyAddress } from 'modules/accountProxy'
+import { useCurrentAccountProxyAddress } from 'modules/accountProxy/hooks/useCurrentAccountProxy'
+import { getProxyAccountUrl } from 'modules/accountProxy/utils/getProxyAccountUrl'
 
 const ProxyPageLink = styled(Link)`
   display: flex;
@@ -36,8 +38,16 @@ export function CowShedInfo({ className, onClick }: CowShedInfoProps): ReactNode
 
   const accountProxyLink = getProxyAccountUrl(chainId)
 
+  const isWidget = isInjectedWidget()
+
   return (
-    <ProxyPageLink to={accountProxyLink} className={className} onClick={onClick}>
+    <ProxyPageLink
+      to={accountProxyLink}
+      className={className}
+      onClick={onClick}
+      target={isWidget ? '_blank' : undefined}
+      rel={isWidget ? 'noopener noreferrer' : undefined}
+    >
       <Pocket size={14} />
       <span>{i18n._(ACCOUNT_PROXY_LABEL)}</span>
     </ProxyPageLink>