+++ apps/cowswap-frontend/src/modules/wallet/containers/Web3Status/index.tsx
@@ -15,9 +15,10 @@ import { Wrapper } from '../../pure/Web3StatusInner/styled'
 export interface Web3StatusProps {
   className?: string
   onClick?: () => void
+  joinedLeft?: boolean
 }
 
-export function Web3Status({ className, onClick }: Web3StatusProps): ReactNode {
+export function Web3Status({ className, onClick, joinedLeft = false }: Web3StatusProps): ReactNode {
   const connectionType = useConnectionType()
   const { account } = useWalletInfo()
   const { ensName } = useWalletDetails()
@@ -27,7 +28,7 @@ export function Web3Status({ className, onClick }: Web3StatusProps): ReactNode {
   const showUnfillableOrdersAlert = useShowUnfillableOrderAlert()
 
   return (
-    <Wrapper className={className} onClick={onClick}>
+    <Wrapper className={className} onClick={onClick} $joinedLeft={joinedLeft}>
       {account && <TradeOrdersPermitUpdater />}
       <Web3StatusInner
         showUnfillableOrdersAlert={showUnfillableOrdersAlert}