+++ apps/cowswap-frontend/src/tradingSdk/tradingSdk.ts
@@ -1,5 +1,9 @@
 import { DEFAULT_APP_CODE } from '@cowprotocol/common-const'
-import { getCurrentChainIdFromUrl, isBarnBackendEnv } from '@cowprotocol/common-utils'
+import {
+  COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS,
+  getCurrentChainIdFromUrl,
+  isBarnBackendEnv,
+} from '@cowprotocol/common-utils'
 import { TradingSdk } from '@cowprotocol/cow-sdk'
 
 import { orderBookApi } from '../cowSdk'
@@ -11,6 +15,7 @@ export const tradingSdk = new TradingSdk(
     chainId,
     appCode: DEFAULT_APP_CODE,
     env: isBarnBackendEnv ? 'staging' : 'prod',
+    settlementContractOverride: COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS,
   },
   {
     orderBookApi,